+++ tests/testthat/test-EKC.R
@@ -20,16 +20,16 @@ test_that("found eigenvalues are correct", {
 })
 
 test_that("reference eigenvalues are correct", {
-  expect_gt(ekc_cor$references[floor(ncol(test_models$baseline$cormat) / 2)], 1)
-  expect_gt(ekc_raw$references[floor(ncol(GRiPS_raw) / 2)], 1)
-  expect_length(ekc_cor$references,
+  expect_equal(ekc_cor$references$BvA2017[floor(ncol(test_models$baseline$cormat) / 2)], 1)
+  expect_equal(ekc_raw$references$BvA2017[floor(ncol(GRiPS_raw) / 2)], 1)
+  expect_length(ekc_cor$references$BvA2017,
                 length(ekc_cor$eigenvalues))
-  expect_length(ekc_raw$references, length(ekc_raw$eigenvalues))
+  expect_length(ekc_raw$references$BvA2017, length(ekc_raw$eigenvalues))
 })
 
 test_that("identified number of factors is correct", {
-  expect_equal(ekc_cor$n_factors, 2)
-  expect_equal(ekc_raw$n_factors, 1)
+  expect_equal(ekc_cor$n_factors_BvA2017, 3)
+  expect_equal(ekc_raw$n_factors_BvA2017, 1)
 })
 
 # Create singular correlation matrix for tests
@@ -52,8 +52,8 @@ test_that("errors are thrown correctly", {
 })
 
 test_that("settings are returned correctly", {
-  expect_named(ekc_cor$settings, c("use", "cor_method", "N"))
-  expect_named(ekc_raw$settings, c("use", "cor_method", "N"))
+  expect_named(ekc_cor$settings, c("use", "cor_method", "N", "type"))
+  expect_named(ekc_raw$settings, c("use", "cor_method", "N", "type"))
 
   expect_equal(ekc_cor$settings$N, 500)
   expect_equal(ekc_raw$settings$N, 810)